moved the controller classes into their own folder
