+++ mcstudio/controllers/EditorCanvasController.cpp
