+++ mcstudio/controllers/WidgetTreeController.cpp
