+++ mcstudio/main.cpp
@@ -18,7 +18,7 @@ int main() {
     auto editor = mc::MakeRef<mc::mcstudio::Editor>();
 
     auto window = mc::mcstudio::McxSourceManager::loadMainWindow();
-    
+
     auto canvas = window->createOverlayCanvas();
     editor->setWindowCanvasReference(canvas);
